KC-7 - Added check kount_session_id for existing, added check for successful login, added check for POST method
